more typecasting

+++ PEAT/utils/gps_test.py
@@ -48,8 +48,8 @@ def convert_to_degrees(raw_value):
 
 def convert_to_meters(lat1, lon1, lat2, lon2):
     R = 6378.137
-    dLat = int(lat2) * math.pi / 180 - int(lat1) * math.pi / 180
-    dLon = int(lon2) * math.pi / 180 - int(lon1) * math.pi / 180
+    dLat = float(lat2) * math.pi / 180 - float(lat1) * math.pi / 180
+    dLon = float(lon2) * math.pi / 180 - float(lon1) * math.pi / 180
     a = math.sin(dLat/2) * math.sin(dLat/2) + math.cos(lat1 * math.pi / 180) * math.cos(lat2 * math.pi / 180) * math.sin(dLon/2) * math.sin(dLon/2)
     c = 2 * math.atan2(math.sqrt(a), math.sqrt(1-a))
     d = R * c
